Added some extra success/failure checking so that if an observation does not get sent to the sequence console, this an error condition is returned.
git-svn-id: file:///jac_sw/svnroot/scicom/trunk/omp/java/qt@4345 64693ad5-8cdc-4012-857f-4cd9bf50b8c8

+++ src/main/edu/jach/qt/gui/ExecuteUKIRT.java
@@ -49,7 +49,7 @@ public class ExecuteUKIRT extends Execute implements Runnable {
 	SpItem itemToExecute;
 	if (!isDeferred) {
 	    itemToExecute = ProgramTree.selectedItem;
-	    logger.info("Executing observatin from Program List");
+	    logger.info("Executing observation from Program List");
 	}
 	else {
 	    itemToExecute = DeferredProgramList.currentItem;
@@ -189,9 +189,13 @@ public class ExecuteUKIRT extends Execute implements Runnable {
 	  //}
 
 	if ( System.getProperty("os.name").equals("SunOS") && TelescopeDataPanel.DRAMA_ENABLED) {
-	  QtTools.loadDramaTasks(inst.type().getReadable());
+	    if (QtTools.loadDramaTasks(inst.type().getReadable()) == 0){
+		failure.delete();
+	    }
+	    else {
+		success.delete();
+	    }
 	}
 	
-	failure.delete();
     }
 }
